extract graph fns to graph_util
