+++ fpr/models/pipeline.py
@@ -3,6 +3,8 @@ import sys
 from dataclasses import dataclass, field
 from typing import AbstractSet, Callable, Optional
 
+from fpr.graph_util import NODE_ID_FORMATS, NODE_LABEL_FORMATS, GROUP_ATTRS
+
 
 def add_infile_and_outfile(
     pipeline_parser: argparse.ArgumentParser
@@ -24,36 +26,6 @@ def add_infile_and_outfile(
         help="pipeline output file (defaults to stdout)",
     )
     return pipeline_parser
-
-
-NODE_ID_FORMATS = {
-    "name": "{pkg_id.name}",
-    "name_version": "{pkg_id.name} {pkg_id.version}",
-    "name_version_source": "{pkg_id.name} {pkg_id.version} {pkg_id.source}",
-    "source": "{pkg_id.source}",
-}
-
-NODE_LABEL_FORMATS = {
-    "name": "{crate.package_id.name}",
-    "name_version": "{crate.package_id.name} {crate.package_id.version}",
-    "name_version_source": "{crate.package_id.name} {crate.package_id.version} {crate.package_id.source}",
-    "source": "{crate.package_id.source}",
-    "name_authors": "{crate.package_id.name}\n{crate_package.authors}",
-    "name_readme": "{crate.package_id.name}\n{crate_package.readme}",
-    "name_repository": "{crate.package_id.name}\n{crate_package.repository}",
-    "name_version_repository": "{crate.package_id.name} {crate.package_id.version}\n{crate_package.repository}",
-    "name_license": "{crate.package_id.name}\n{crate_package.license}",
-    "name_package_source": "{crate.package_id.name}\n{crate_package.source}",
-    "name_metadata": "{crate.package_id.name}\n{crate_package.metadata}",
-}
-
-GROUP_ATTRS = {
-    "author": lambda node: node[1]["crate_package"].authors or [],
-    "repository": lambda node: node[1]["crate_package"].repository or "",
-    # 'workspace':
-    # 'manifest_path':
-    # 'source_repository':
-}
 
 
 def add_graphviz_graph_args(parser: argparse.ArgumentParser) -> argparse.ArgumentParser:
